+++ business_rules/javascript/business_rules_TijsDeBelie/tests/support/jasmine.json
@@ -1,10 +1,7 @@
 {
-  "spec_dir": "spec",
+  "spec_dir": "tests",
   "spec_files": [
     "**/*[sS]pec.js"
-  ],
-  "helpers": [
-    "helpers/**/*.js"
   ],
   "stopSpecOnExpectationFailure": false,
   "random": true
